Pass a transaction into Postgres
If we start a job, then the power cord of the computer is kicked out of
the wall before we complete it, the expectation is that the job will be
handed to another worker. If we don't pass a transaction, the row
deletion for the taken job is deleted immediately and the job will never
be retried.